feat: signed prices — Price becomes i64 for negative-price markets
Price (and tick_size) switch u64 -> i64; tick stays strictly positive
(tick_size <= 0 rejected). Positive canonical bytes are unchanged
(8-byte LE), so no migration. Gossip gains an i64 Reader; book/entry
math goes through i128. Margin estimates pick the larger *magnitude*
(abs max), overflow guard uses unsigned_abs, and the +/-10% mark cap
plus slash-deviation now measure |old| so negative marks move by
capped steps instead of freezing. Fill mark gate charges |notional|.
watch prove.rs parses signed fill prices; fees charge |notional|.
Tests: negative short PnL sign, negative-mark snapshot margins,
capped negative mark stepping, spot_only x negative fill joint
regression; e2e gains a negative-price fill_math fraud case.

+++ crates/operp-exec/examples/bench_raw.rs
@@ -42,7 +42,7 @@ fn main() {
         tip = unit_id(&u);
         eng.ingest(u).unwrap();
     }
-    let px = 100_000 * PRICE_SCALE;
+    let px = 100_000 * PRICE_SCALE as i64;
     let qty = QTY_SCALE / 100;
     let seqs = std::cell::RefCell::new([1u64; 4]);
     // single-threaded open/close pairs on one market, direct ingest
